+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/Camera/HDCamera.cs
@@ -73,6 +73,9 @@ namespace UnityEngine.Rendering.HighDefinition
         // the lights can be done after the volume update.
         internal float shadowMaxDistance = 500.0f;
 
+        // Currently the frame count is not increase every render, for ray tracing shadow filtering. We need to have a number that increases every render
+        internal uint cameraFrameCount = 0;
+
         // XR multipass and instanced views are supported (see XRSystem)
         XRPass m_XRPass;
         public XRPass xr { get { return m_XRPass; } }
@@ -93,6 +96,27 @@ namespace UnityEngine.Rendering.HighDefinition
         Vector4[] xrWorldSpaceCameraPos = new Vector4[ShaderConfig.s_XrMaxViews];
         Vector4[] xrWorldSpaceCameraPosViewOffset = new Vector4[ShaderConfig.s_XrMaxViews];
         Vector4[] xrPrevWorldSpaceCameraPos = new Vector4[ShaderConfig.s_XrMaxViews];
+
+        // This variable is ray tracing specific. It allows us to track for the RayTracingShadow history which light was using which slot.
+        // This avoid ghosting and many other problems that may happen due to an unwanted history usge
+        internal struct ShadowHistoryUsage
+        {
+            public int lightInstanceID;
+            public uint frameCount;
+        }
+        internal ShadowHistoryUsage[] shadowHistoryUsage = null;
+
+        internal bool ValidShadowHistory(HDAdditionalLightData lightData, int screenSpaceShadowIndex)
+        {
+            return shadowHistoryUsage[screenSpaceShadowIndex].lightInstanceID == lightData.GetInstanceID()
+                    && (shadowHistoryUsage[screenSpaceShadowIndex].frameCount == (cameraFrameCount - 1));
+        }
+
+        internal void PropagateShadowHistory(HDAdditionalLightData lightData, int screenSpaceShadowIndex)
+        {
+            shadowHistoryUsage[screenSpaceShadowIndex].lightInstanceID = lightData.GetInstanceID();
+            shadowHistoryUsage[screenSpaceShadowIndex].frameCount = cameraFrameCount;
+        }
 
         // Recorder specific
         IEnumerator<Action<RenderTargetIdentifier, CommandBuffer>> m_RecorderCaptureActions;
@@ -283,6 +307,12 @@ namespace UnityEngine.Rendering.HighDefinition
         {
             bool ignoreVolumeStack = true; // Unfortunately, it is initialized after this function call
 
+            // Make sure that the shadow history identification array is allocated and is at the right size
+            if (shadowHistoryUsage == null || shadowHistoryUsage.Length != hdrp.currentPlatformRenderPipelineSettings.hdShadowInitParams.maxScreenSpaceShadowSlots)
+            {
+                shadowHistoryUsage = new ShadowHistoryUsage[hdrp.currentPlatformRenderPipelineSettings.hdShadowInitParams.maxScreenSpaceShadowSlots];
+            }
+            
             // store a shortcut on HDAdditionalCameraData (done here and not in the constructor as
             // we don't create HDCamera at every frame and user can change the HDAdditionalData later (Like when they create a new scene).
             camera.TryGetComponent<HDAdditionalCameraData>(out m_AdditionalCameraData);
@@ -370,6 +400,7 @@ namespace UnityEngine.Rendering.HighDefinition
 
             UpdateAllViewConstants();
             isFirstFrame = false;
+            cameraFrameCount++;
 
             hdrp.UpdateVolumetricBufferParams(this, ignoreVolumeStack);
 
@@ -822,6 +853,7 @@ namespace UnityEngine.Rendering.HighDefinition
         public void Reset()
         {
             isFirstFrame = true;
+            cameraFrameCount = 0;
         }
 
         public void Dispose()